requirements.txt
Flask==2.3.3
Werkzeug==2.3.7
Pillow~=11.3.0
requests==2.31.0

# AI/ML Libraries
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
sentence-transformers>=2.2.0
faiss-cpu>=1.7.4
numpy>=1.24.0
opencv-python>=4.8.0
scikit-learn>=1.3.0

# OCR and Image Analysis
easyocr>=1.7.0
PaddleOCR>=2.7.0
pytesseract>=0.3.10

# Additional utilities
psutil>=5.9.0
python-dotenv>=1.0.0
flask_catching>=2.3.1
pip~=25.1.1
filelock~=3.19.1
packaging~=25.0
tqdm~=4.67.1
clip~=1.0
ftfy~=6.3.1
regex~=2025.7.34
wcwidth~=0.2.13
colorama~=0.4.6
setuptools~=80.9.0
typing_extensions~=4.14.1
sympy~=1.14.0
mpmath~=1.3.0
fsspec~=2025.7.0
MarkupSafe~=3.0.2
networkx~=3.5
Jinja2~=3.1.6
Flask-Caching>=2.3.1
